+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [2.18.0] - 2021-06-21
+
 ## Added
 
 - Add releasemaxhistory resource which ensures we retry at a reduced rate when
@@ -366,7 +368,8 @@ We check progress in the next reconciliation loop. ([#362](https://github.com/gi
 - Separate podsecuritypolicy.
 - Security context in deployment spec with non-root user.
 
-[Unreleased]: https://github.com/giantswarm/chart-operator/compare/v2.17.0...HEAD
+[Unreleased]: https://github.com/giantswarm/chart-operator/compare/v2.18.0...HEAD
+[2.18.0]: https://github.com/giantswarm/chart-operator/compare/v2.17.0...v2.18.0
 [2.17.0]: https://github.com/giantswarm/chart-operator/compare/v2.16.0...v2.17.0
 [2.16.0]: https://github.com/giantswarm/chart-operator/compare/v2.15.0...v2.16.0
 [2.15.0]: https://github.com/giantswarm/chart-operator/compare/v2.14.0...v2.15.0